fix(rm): 64-bit dedicated server RUNS a map with native qagame module
Three classic id Tech 3 64-bit bugs fixed (found via the new trace log) -
the heap corruption that crashed init is gone, and etrmded now boots, loads
2.60b paks, loads a BSP, runs GAME_INIT + sim frames in the 64-bit
qagame_mp_x86_64.dll, and stays up. Verified on oasis and radar.

Bugs:
- files.c FS_AddGameDirectory: qsort element size hardcoded 4 -> sizeof(char*)
  (pointer array; half-pointer swaps corrupted the pak list).
- win_main.c Sys_ListFiles/Sys_ListFilteredFiles: _findfirst handle stored in
  int -> intptr_t (pointer-sized; truncation corrupted the CRT find state).
- common.c Com_InitHunkMemory: hunk base cacheline-align cast (int)s_hunkData
  -> (intptr_t) (truncated the 64-bit malloc base to a low garbage address ->
  every Hunk_Alloc returned a bad pointer -> heap corruption). THE big one.

Also adds Com_RMTrace(): a crash-proof, env-gated (ETRM_TRACE=<path>) trace
logger that flushes per line, plus dormant trace points through Com_Init and
SV_SpawnServer. Off by default; invaluable for pre-console-subsystem debugging.

+++ src/win32/win_main.c
@@ -255,7 +255,7 @@ DIRECTORY SCANNING
 void Sys_ListFilteredFiles( const char *basedir, char *subdirs, char *filter, char **list, int *numfiles ) {
 	char search[MAX_OSPATH], newsubdirs[MAX_OSPATH];
 	char filename[MAX_OSPATH];
-	int findhandle;
+	intptr_t findhandle;   // 64-bit: _findfirst handle is pointer-sized
 	struct _finddata_t findinfo;
 
 	if ( *numfiles >= MAX_FOUND_FILES - 1 ) {
@@ -325,7 +325,7 @@ char **Sys_ListFiles( const char *directory, const char *extension, char *filter
 	char        **listCopy;
 	char        *list[MAX_FOUND_FILES];
 	struct _finddata_t findinfo;
-	int findhandle;
+	intptr_t findhandle;   // 64-bit: _findfirst/_findnext use pointer-sized handles (was int -> truncation -> CRT crash)
 	int flag;
 	int i;
